crimson/net: introduce protocol-level abstraction
